services/web/requirements.txt
Flask-SQLAlchemy==3.0.3
gunicorn
psycopg2-binary==2.9.6
astroid
requests
Babel
black
click
flake8
Flask-Babel
Flask-Cors
Flask-Login
Flask-Table
greenlet
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
mypy-extensions
pathspec
pi
platformdirs
pycodestyle
pyflakes
pylint
pytz
regex
six
SQLAlchemy
toml
tomli
typing-extensions
Werkzeug
wrapt
WTForms
Flask-Migrate
pandas
python-dotenv